Moved create experiment run code to seperate handler
